Updated part3.cpp include <math>
Began work on prime factors greater than 2, and included the math liubrary in order to check the square root of a given value. As the definition of composite numbers and their prime factors states that a prime factor will be less than or equal to the square root of the given value.